requirements.txt
pandas
pdfplumber
google-generativeai
python-dotenv
openpyxl
streamlit
